+++ QuanLyThuVien/Controllers/HomeController.cs
@@ -1,12 +1,23 @@
-﻿using System;
+﻿using QuanLyThuVien.DAL;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Web;
 using System.Web.Mvc;
+using QuanLyThuVien.Models;
 
 namespace QuanLyThuVien.Controllers {
     public class HomeController : Controller {
+
+        private LibraryContext db = new LibraryContext();
+
+
         public ActionResult Index() {
+            if (Session["ID"] != null) {
+                var id = Int32.Parse(Session["ID"].ToString());
+                Librarian user = db.Librarians.Find(id);
+                ViewBag.Librarian = user;
+            }
             return View();
         }
 
@@ -21,5 +32,23 @@ namespace QuanLyThuVien.Controllers {
 
             return View();
         }
+
+        public ActionResult Login() {
+            ViewBag.Message = "Login";
+            return View();
+        }
+
+        [HttpPost]
+        public ActionResult CheckAccount(string uid, string password)
+        {
+            /* 
+             This try catch to catch external error of system when data is indexing
+             */
+            var librarians = db.Librarians.ToList();
+
+            return View();
+        }
+
+
     }
 }