+++ QuanLyThuVien/Models/Librarian.cs
@@ -1,0 +1,39 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+using System.ComponentModel.DataAnnotations;
+
+namespace QuanLyThuVien.Models {
+    public class Librarian {
+        [Key]
+        public int ID { get; set; }
+
+        [Required(ErrorMessage ="UID is required.")]
+        public string UID { get; set; }
+
+        [Required(ErrorMessage ="Password is required.")]
+        [DataType(DataType.Password)]
+        public string Password { get; set; }
+
+        [Required(ErrorMessage ="Please confirm password.")]
+        [DataType(DataType.Password)]
+        public string ConfirmPassword { get; set; }
+
+        [Required(ErrorMessage ="First Name is required.")]
+        public string FirstName { get; set; }
+
+        [Required(ErrorMessage ="Last Name is required.")]
+        public string LastName { get; set; }
+
+        [Required(ErrorMessage ="Email is required.")]
+        [EmailAddress]
+        public string Email { get; set; }
+
+        public string Gender { get; set; }
+
+        public string Address { get; set; }
+
+        public string Phone { get; set; }
+    }
+}