+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 [0.1] 16/04/2016
 
 ## Added
+- Librarian Controller, View, Model
+- Layout
+- Login, Register, Edit Librarian
+
+[0.1] 16/04/2016
+
+## Added
 - CHANGELOG.md file